fix some changes to complete merge
- TagsList component used in metric (create/edit) as tag names list.
- TagsObj component used in service (create/edit) as service tag obj list {id, name}
