+++ frontend/src/routes/MainRouter.jsx
@@ -1,12 +1,22 @@
 import React from "react";
-import { createBrowserRouter } from "react-router-dom";
+import { createBrowserRouter, Navigate } from "react-router-dom";
 
-// All Pages
+// Metric Pages
 import Metrics from '../pages/Metrics';
 import MetricCreateEdit from '../pages/MetricCreateEdit';
+// Services Pages
+import Services from "../pages/Services";
+import ServiceCreateEdit from "../pages/ServiceCreateEdit";
+import ServiceInfo from "../pages/ServiceInfo";
+import ScorecardMetrics from "../pages/ScorecardMetrics";
+// NotFound
 import NotFound from '../pages/NotFound';
 
 const MainRouter = createBrowserRouter([
+  {
+    path: "/",
+    element: <Navigate to="/dashboard/services" replace/>
+  },
   {
     path: "/dashboard",
     children: [
@@ -23,6 +33,27 @@ const MainRouter = createBrowserRouter([
         path: "metrics/edit/:metric_id",
         element: <MetricCreateEdit mode="edit"/>
       },
+      // Service routes
+      {
+        path: "services",
+        element: <Services />,
+      },
+      {
+          path: "services/:service_id",
+          element: <ServiceInfo />
+      },
+      {
+          path: "services/:service_id/:scorecard_id",
+          element: <ScorecardMetrics />
+      },
+      {
+        path: "services/create",
+        element: <ServiceCreateEdit mode="create"/>
+      },
+      {
+        path: "services/edit/:service_id",
+        element: <ServiceCreateEdit mode="edit"/>
+      }
     ]
   },
   // Not Found
